Optimization of wordsNearest
After performing the multiplication, sorting the whole matrix and then extracting the top N words is expensive. I switched the implementation to a priority queue. The result is a boost in the performance. For the word `France` in the `GoogleNews` corpus, the original implementation takes `2908 ms` in sorting. The new implementation takes only `102 ms` to perform the extraction.

Former-commit-id: 1118d3d82adbbfba9ced25643dcc52b7033b610c [formerly 0ec489c119cd004ec26434796ee984781188b894]
Former-commit-id: 9aa2df33fbad713f443acd7949ab839cadf188d8