Runtime Casting Errors with Complex NDArrays
A number of unit tests fail with runtime casting errors
between complex and real-valued NDArrays.

The problem is that BaseNDArray sometimes calls Nd4j.create
directly. This returns a real-valued NDArray. When derived
classes are complex-valued, this leads to a runtime error.

The fix is to put all calls to Nd4j.create inside create wrappers
and call either Nd4j.create or Nd4j.createComplex as
appropriate.