Add epsilon INDArray as a member of MultiLayerNetwork
This will be updated after calls to backprop(), which previously uses
the first projection of the output from calcBackpropGradients() to
update the top-level gradient, but does not use the second projection
(the epsilons) for anything.

A getter for epsilon is also implemented.

This closes #1201; my initial requirement was to obtain the
backpropagated errors at the neural network's input.


Former-commit-id: 01e4713936bf2691db63fc543cad3b43468c5a04 [formerly 3688b7d475c32eebfd0866b15768d151c4668fbc]
Former-commit-id: bee6412aa0cb8261f2c7b52ab860a9d1cb45ede6