LearningRate and BiasLearningRate
The default behavior of Learning and BiasLearningRate is error-prone if you intend to change the learning rate of one specific layer.
In the current case, if you do, you most likely want to specify the same rate for the bias learning rate. Here unless otherwise specified it will apply the global learningRate instead of applying the learningRate of that specific layer.