LearningRate and BiasLearningRate (#1412)
The default behavior of Learning and BiasLearningRate is error-prone if you intend to change the learning rate of one specific layer.
In the current case, if you do, you most likely want to specify the same rate for the bias learning rate. Here unless otherwise specified it will apply the global learningRate instead of applying the learningRate of that specific layer.

Former-commit-id: f3886a94e9c8cd5974b9d5749d7162ebe3dda948 [formerly d5bdbaf36271d81b39322c3b784c751705f423dc]
Former-commit-id: f9a467ee0e4921faf9e6de9dcce7ae0f1066d3f9