Revert "Fix wrong variable expansion"
This reverts commit 1af297967e0a45ed44489ead826695121037aad3.

Can't find a way to make this work across the board. As Windows
doesn't seem to care that much about RPATH anyway, I'll simply
revert this change.

Note: On Linux the expansion is still wrong on the first calls
to g++ but when the linker is invoked it is magically right again.
So something is still problematic here, but it ends up in the
correct behavior in the place where it matters. Using the
recommended way of setting CMAKE_INSTALL_RPATH, I still couldn't
make it work.