Add single file serializer/unserializer for normalization pre-processors
When saving a NormalizerStandardize instance, one had to pass 2 or 4 files depending on whether labels were fit.
With the new MultiNormalizerStandardize, that was 2 * inputs + 2 * outputs, which can become quite unwieldly.

This new utility can serializer and unserialize normalizers to/from self contained ZIP archives. As a bonus,
it also saves settings like whether or not to normalizer labels.