Add TransformSplit class to map one InputSplit to another one
I have a use case where I need my inputs and outputs in separate files because I'm dealing with time series
and outputs of different lengths. There was no way to pair two FileSplits into a single iterator because
the order of the files is not deterministic.

This utility solves the problem by transforming the URIs of one split into different URIs. Based on any custom
naming scheme, one can then create a split of output files based on the input files in the same order, so it
even works when the input split it shuffled.