Fix vertex ordering issue when cloning computation graph config (#2436)
When using `SparkComputationGraph.calculateScoreMultiDataSet()` I was getting different scores than when not using Spark.
I found out that the problem was that the `clone()` function in `ComputationGraphConfiguration` class used `HashMap` instead
of `LinkedHashMap`, resulting in a random ordering of outputs when JSON serializing the configuration. Effectively, the weights
for outputs were being assigned to the wrong outputs when restoring the network, resulting in wrong scores.

This change fixes it.