Make 'mvn compile' great again
TL;DR : `mvn compile` (and `test`) fails because the shade plugin
structurally can't make the classes it's supposed to shade available
before the `package` phase, if you stay within the normal maven
lifecycle.

The soltion implemented here is to use the
[jetspeed-mvn](https://portals.apache.org/jetspeed-2/buildguide/jetspeed-mvn-plugin.html)
plugin in the shaded project, bound to phase `compile`, to create a
set of invocations that fall outside of the normal build lifecycle.
TL;DR : for the `nd4j-shade/jackson project` (and it alone), 'compile'
now means 'compile + package'.

A caveat about this is that phases before `package` in projects that
*depend* on `org.nd4j.shade.*` packages will not expect to find
sources packaged in the form of a jar in the `nd4j-shade/jackson`
project, but they will rather expect to find unpacked classes (the way
maven creates them in all other of their dependencies). But our
triggered execution of `package` in `nd4j-shade/jackson` creates a
jar, leaving no classes behind. To put them in place, we just have to
re-unpack them using the [maven-dependency
plugin](https://maven.apache.org/plugins/maven-dependency-plugin/).

Fixes #1658