Add/restore normalizers to/from model files using new normalizer serializer (#2957)
Instead of using Java serialization, this diff uses the new NormalizerSerializer that serializes to a custom binary format.
In order to support model files with normalizers added with the old functionality, the deprecated unserialization function is
kept and used as a fallback approach. If an old format is detected, it will automatically be fixed by re-saving the normalizer.

Former-commit-id: 5418e1e856b2a805514cf1c288f43664a72fd50c