Add/restore normalizers to/from model files using new normalizer serializer (#2957)
Instead of using Java serialization, this diff uses the new NormalizerSerializer that serializes to a custom binary format.
In order to support model files with normalizers added with the old functionality, the deprecated unserialization function is
kept and used as a fallback approach. If an old format is detected, it will automatically be fixed by re-saving the normalizer.

Former-commit-id: 13501e244cec35b5046a739cf6cab7f4f4f75874 [formerly 5418e1e856b2a805514cf1c288f43664a72fd50c]
Former-commit-id: 5d49c4b755b9be70e1c5b40d7ab6171aa528f1b0