Revert "Force cleaning of sources on spark-x modules"
This reverts commit cfbeb3b921e8a10d829f142f674cf4f5129e1ccc.

This erased the jar created in a normal 'mvn install' run.
This revert will also create issues in ./buildmultiplescalaversions.sh
runs, where the non-cleaning of tasks will make a spark-2 build pick up a
spark-1 artefact.

The fix would be to 'mvn clean' between changes of versions,
but then we run into #2873

Hotfixing this for now, as ./buildmultiplescalaversions.sh runs
are less frequent.