Provides an implementation of multi-probe LSH for the cosine distance.
This can be used as guidelines for a poor man's ANN if the distribution of
the data is well-known enough to be able to predict an average distance
 between ANNs (the radius parameter of Panigraphi's entropy LSH).

However, because if uses a single hash table (as opposed to the several of the
vanilla LSH impelementation), it is relatively reasonable in its memory usage.

In the case of a very large index that would have to be split on several machins
the flavor of LSH that should be used is Layered LSH ( https://arxiv.org/abs/1210.7057 )

WIP, in particular the index creation does not use workspaces for now.

Opened as a request for comments (hi @turambar!) and perhaps an inspiration
to upgrade this to another implementation (hi @turambar!).

From the discard pile of high-performance ANN experiments.