Add coefficient of determination metric
Distinguish between Pearson Correlation and Coefficient of Determination
as the implementation of 'correlationR2' actually calculated the
Pearson Correlation.
The 'correlationR2' method still returns the Pearson Correlation to
reduce the confusion of users. Though, it is now deprecated and calls
the
'pearsonCorrelation' method.
Additionally the 'rSquared' method was added, which calculates the
Coefficient of Determination.
Test values were added as well and were computed with
sklearn.metrics.r2_score

Fixes: #4342