Resolve thread leaks by switching to threadly
This resolves #4766 as well as a couple other thread leaks I found.
Threadly pools perform great, are robust (often times with greater garuntees than java.util), and in this case wont leak threads if garbage collected.
Not all cases updated in this commit were leaks, but I did update several places to maintain consistency in the project.  I think in the future we may want to make the lifecycle of these schedulers a bit cleaner, but this provides an immediate win.
`ExecutorService` was updated to the more generic `Executor` where the executor is only needed for execution (and lifecycle is not managed)

By default threadly schedulers are daemon threads, so in many cases the code was simplified as well.
Threadly's version at the time of this commit is 5.14, which does offer further performance gains and features.  But does require java 8.  So this commit depends on 4.10.0 which works with java 6+.