Frozen layer with backprop implementation
use backpropGradient method of inside layer, but make sure
the graidents are left at zero so that the parameters
are not changed. (Same effect as using SGD updater with 0.0
learning rate).

Also added test for instantiation, check if frozen layers changed
and compare the whole network with same network using Sgd updater.
Done for both multi layer network and computation graph.

This should solve #4760 as well