Handle stdev = 0 case
In forward pass 0/0 = 0 is good enough. In backward pass, gradient in
that point is undefined as lim x->-0 = -inf while lim x-> +0 = inf.
Instead of allowing inf or NaN to propagate, we set the gradient to 0 in
that case, as it should provide a more reasonable overall gradient in
the common case.