#10154: Redirected all basic writes to the DataOutputStream
This is done to ensure that the underlying stream is notified when it's closed by any of the write calls (more specifically the NormalizerSerializer.getDefault().write). Previously, the NormalizerSerializer.getDefault().write would close the zipfile stream, but since that's wrapped in the DataOutputStream, the dos would not be notified of the closed stream, and would fail on the dos.close() on line 194.

Signed-off-by: Siebe Krijgsman <skrijgsman@gmail.com>